refactored shuffledModel CP as a helper and updated tests and imports
