+++ app/people/controller.js
@@ -1,5 +1,6 @@
 import Controller from '@ember/controller';
 import {computed} from '@ember/object';
+import {shuffle} from 'star-wars-trumps/utils/helpers';
 
 export default Controller.extend({
   leftPlayerScore: 0,
@@ -8,10 +9,6 @@ export default Controller.extend({
   shuffledModel: computed('model', function () {
     const array = this.model.toArray();
 
-    for (let i = array.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [array[i], array[j]] = [array[j], array[i]];
-    }
-    return array;
+    return shuffle(array);
   })
 });
