--- app/utils/random-integer.js
@@ -1,5 +1,0 @@
-export default function getRandomInt(min, max) {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
